feat: Command Center tab system with inline agent management
## New Components
- AgentCreate: Reusable multi-step agent creation form
- AgentDetail: Added defaultTab prop for auto-selecting tabs

## Command Center Agents Tab
- Clicking agent card shows AgentDetail inline (no navigation)
- Clicking 'Create Agent' shows AgentCreate inline
- Tab state resets when switching away from Agents tab

## Sidebar Improvements (MemberList)
- 'View Profile' → Shows agent detail (Activity tab)
- 'Settings' → Shows agent detail (Settings tab auto-selected)
- 'Add Agent' → Shows create form inline
- All actions stay within Command Center (no navigation)

## Refactored Pages
- /agents/create/page.tsx now uses AgentCreate component
- /agents/[agentId]/page.tsx uses AgentDetail component
